+++ src/app/api/portfolio/apps/[pid]/route.ts
@@ -3,9 +3,9 @@ import prisma from '@/utils/prisma';
 import fs from 'fs';
 import { join } from 'path';
 import { nanoid } from 'nanoid';
-import { rmFile } from 'rm-file';
 import getFieldsFromForm from '@/utils/getFieldsFromForm';
 import mime from 'mime-types';
+import getUser from '@/utils/getUser';
 
 const uploadPath = './public/uploads/';
 
@@ -15,6 +15,12 @@ export async function PUT(
     params: Promise<{ pid: string }>;
   },
 ) {
+  const user = await getUser();
+  if (!user) {
+    return new Response('Unauthorized', {
+      status: 401,
+    });
+  }
   const params = await props.params;
   const { pid } = params;
   const appReturn = await prisma.app.findFirst({
@@ -74,6 +80,12 @@ export async function DELETE(
     params: Promise<{ pid: string }>;
   },
 ) {
+  const user = await getUser();
+  if (!user) {
+    return new Response('Unauthorized', {
+      status: 401,
+    });
+  }
   const params = await props.params;
   const { pid } = params;
   await prisma.app.delete({
